✨ feat(重要token识别): 初步完成功能实现
初步完成了使用重要token的推理过程，手造两个输入结果正确。具体正确情况需要进一步修改验证
代码由于前后思路对于是否要batching的思路不统一，需要修改
prefill的计时器可能把持久化前缀kv的时间算上了，需要调整
部分代码逻辑较凌乱，需要整理

+++ flexllmgen/kv_cache_manage.py
@@ -185,8 +185,8 @@ class Chunk:
         self.chunk_id = chunk_id or Chunk.next_chunk_name()
         # shape = (1, batch*n_probe_head, head_dim)
         # shape = (chunk_size, batch * n_heads, head_dim)
-        self.full_head_k = device.allocate(shape=full_head_shape, dtype=np.float32, pin_memory=True)
-        self.full_head_v = device.allocate(shape=full_head_shape, dtype=np.float32, pin_memory=True)
+        self.full_head_k = device.allocate(shape=full_head_shape, dtype=np.float16, pin_memory=True)
+        self.full_head_v = device.allocate(shape=full_head_shape, dtype=np.float16, pin_memory=True)
     
     @classmethod
     def next_chunk_name(cls):
